+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Internal builds may append content to the Unreleased section.
 Only write entries that are worth mentioning to users.
 -->
 
+## [Unreleased]
+
+- Lib: Support creating named sessions via `Session.create(work_dir, session_id)`
+- CLI: Automatically create new session when specified session ID is not found
+
 ## [0.64] - 2025-12-15
 
 - UI: Fix UTF-16 surrogate characters input on Windows